# Speed up bottleneck in some differential unit lookup