Representation objects store differential (velocity) data in a dictionary `.differentials` with keys set to the SI unit with which the differential is take. So, for velocities, this is `'s'`, because velocities are time derivatives of positional information.

This PR makes a small change to the way `BaseDifferential._get_deriv_key()` gets the differentials dictionary key from specified differential and representation objects. For scalar input, this makes `_get_deriv_key()` ~20x faster, which makes `representation.with_differentials()` ~16x faster, which makes frame initialization ~7x faster (which also propagates to some paths through the SkyCoord initializer) when there is velocity data.

Defining:
```python
rep = coord.CartesianRepresentation([1., 2, 3] * u.kpc)
dif = coord.CartesianDifferential([1, 2, 3.] * u.km/u.s)
```

This PR:
```python
%timeit dif._get_deriv_key(rep)
238 µs ± 18.2 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
%timeit rep.with_differentials(dif)
662 µs ± 58.6 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
%timeit coord.ICRS(ra=1*u.deg, dec=2*u.deg, pm_ra_cosdec=1*u.mas/u.yr, pm_dec=2*u.mas/u.yr)
1.23 ms ± 140 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```

Master:
```python
%timeit dif._get_deriv_key(rep)
5.41 ms ± 123 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
%timeit rep.with_differentials(dif)
10.8 ms ± 329 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
%timeit coord.ICRS(ra=1*u.deg, dec=2*u.deg, pm_ra_cosdec=1*u.mas/u.yr, pm_dec=2*u.mas/u.yr)
7.22 ms ± 969 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```