@pllim 2018-10-19
Add a benchmark to https://github.com/astropy/astropy-benchmarks/blob/master/benchmarks/coordinates.py ?