@adrn 2018-10-19
See astropy/astropy-benchmarks#68 for benchmarks!