@eteq 2018-10-19
LGTM, so merging so as not to block some other planned coordinates enhancements.  

As a side-note, we may want to re-work some of these preformance enhancements notes into a dedicated section, hence the "what's new needed" label. But not the end of the world if we don't get to that by release time.